kalakitchen/requirements.txt
# KalaKitchen ML Workflow Dependencies
google-generativeai>=0.3.0
opencv-python>=4.8.0
whisper>=1.1.10
pytesseract>=0.3.10
requests>=2.31.0
beautifulsoup4>=4.12.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
aiofiles>=23.0.0
httpx>=0.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
youtube-dl>=2021.12.17
ffmpeg-python>=0.2.0